Lots of changes
I've changed the extention to .php to enable sql request and wrote a small to do list at the top in order to explain how to fill the different parts. 
Like we discussed there are two main parts that i've named "pointer on the map part" and "template part".
So for the first part there's an explanation on how to edit the geojson file and then the creation of an external .php file in order to partition the main code to keep it manageable.
Same thing for the second part that uses sql requests and the database.
Please take a look at it and maybe try the steps to tell me if they are understandable because i personally thought that it was very confusing while re-reading it lol.